Extract `toTemplateData` out into its own file
